Updated Web Page
Added the images "Southern-California-Beaches.jpg" and "WebDesign.jpg"
to the images folder.

Updated the first image in the header to be WebDesign.jpg and updated
the alt content. Also, updated the background image in the header via
CSS as well as applying "transform: scale(1.2) rotate(-365deg);" to the
.profile-image:hover in the styles.css file.

Finally, in the index.html page I updated the ".skills" ul to contain
the skills that I want to learn and share what I'm most interested in.